Register Controller Bug Fixing

+++ src/utils/cloudinary.js
@@ -1,42 +1,36 @@
-import {v2 as cloudinary} from cloudinary
-import fs from fs
+import dotenv from "dotenv";
+dotenv.config();
 
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
- cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY, 
-        api_secret:  process.env.CLOUDINARY_API_SECRET
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
+const uploadOnCloudinary = async (localFilePath) => {
+    try {
+        if (!localFilePath) return null;
 
-const uploadOnCloudinary =async (localFilePath)=>{
-      try{
-          if(!localFilePath)  return null;
+        //upload file on the cloudinary
 
-          //upload file on the cloudinary
-         
-        const uploadResult = await cloudinary.uploader
-       .upload(
-            localFilePath, {
-               resource_type : "auto",
-           }
-       )
-       
+        const uploadResult = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto",
+        });
+
+        console.log("File uploaded successfully !");
+        fs.unlinkSync(localFilePath)
+
         return uploadResult;
-          
-        console.log("File uploaded successfully !")
-        console.log(uploadResult);
-        console.log(uploadResult.url);
+    } catch (err) {
+        console.log(err);
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath); //remove the locally save temporary file as upload operation failed
+        }
+        return null;
+    }
+};
 
-      }
-      catch(err)
-      {
-
-          console.log(err)
-          fs.unlinkSync(localFilePath) //remove the locally save temporary file as upload operation failed
-          return null
-      }
-    
-}
-
-export {uploadOnCloudinary}
+export { uploadOnCloudinary };
